This is a minor improvement to using `relevel` on `annotated_factor`s…

+++ R/annotated_data.R
@@ -744,7 +744,7 @@ setMethod(
 #' @seealso \code{\link{relevel}}
 #' @details Currently, two methods for reordering the levels of an \code{annotated_factor} are implemented:
 #' The \code{annotated_factor,character}-method reorders the levels according to the ordering in a character vector \code{ref}.
-#' The \code{annotated_factor,integer}-method reorders the levels according to their position within the vector of levels
+#' The \code{annotated_factor,numeric}-method reorders the levels according to their position within the vector of levels
 #' (as would be returned by \code{levels}).
 #'
 #' @rdname relevel_annotated
@@ -776,7 +776,7 @@ setMethod(
 
 setMethod(
   f = "relevel"
-  , signature = c(x = "annotated_factor", ref = "integer")
+  , signature = c(x = "annotated_factor", ref = "numeric")
   , definition = function(x, ref){
     # reorder integer vector:
     index <- 1:length(x@levels)
